Extract x86_64 dispatcher initialization to arch
