+++ framework/aster-frame/src/device/dispatcher/io_mem.rs
@@ -8,7 +8,7 @@ use id_alloc::IdAlloc;
 use log::{debug, info};
 
 use crate::{
-    boot::memory_region::MemoryRegion, device::io_mem::IoMem, sync::SpinLock, vm::PAGE_SIZE,
+    arch::dispatcher::init_io_mem_dispatcher, device::io_mem::IoMem, sync::SpinLock, vm::PAGE_SIZE
 };
 
 struct IoMemAllocator {
@@ -112,7 +112,7 @@ impl IoMemDispatcher {
     /// # Safety
     ///
     /// User must ensure the range doesn't belong to physical memory.
-    unsafe fn add_range(&self, range: Range<usize>) {
+    pub(crate) unsafe fn add_range(&self, range: Range<usize>) {
         let start_addr = range.start.align_up(PAGE_SIZE);
         let end_addr = range.end.align_down(PAGE_SIZE);
         info!("Adding MMIO range:{:x}...{:x}", start_addr, end_addr);
@@ -132,61 +132,5 @@ pub static IO_MEM_DISPATCHER: IoMemDispatcher = IoMemDispatcher::new();
 
 pub(crate) fn init() {
     // Initialize the allocator of the IoMemDispatcher based on the system memory info.
-    #[cfg(target_arch = "x86_64")]
-    init_dispatcher(crate::boot::memory_regions().to_owned());
-}
-
-/// This function will initialize the allocatable MMIO area based on the x86-64 memory distribution map.
-///
-/// In x86-64, the available physical memory area is divided into two regions below 32 bits (Low memory)
-/// and above (High memory). Where the area from the top of low memory to 0xffff_ffff, after the top of
-/// high memory, is the available MMIO area.
-///
-#[cfg(target_arch = "x86_64")]
-fn init_dispatcher(mut regions: Vec<MemoryRegion>) {
-    use crate::boot::memory_region::MemoryRegionType;
-
-    if regions.is_empty() {
-        panic!("Memory Regions is empty.")
-    }
-    regions.sort_by_key(|a| a.base());
-    regions.reverse();
-
-    let mut is_tohm_found = false;
-    const HIGH_MMIO_SIZE: usize = 0x100_0000;
-    for region in regions {
-        if !is_tohm_found && (region.base() + region.len()) > u32::MAX as usize {
-            // Find the TOHM (Top of High Memory) and initialize High MMIO region.
-            // Here, using 16MB (0x100_0000) as the size of High MMIO region.
-            // TODO: Update the High MMIO region in runtime.
-            if region.typ() == MemoryRegionType::Usable {
-                is_tohm_found = true;
-                // Align start address to 0x8_0000_0000
-                let mmio_start_addr = (region.base() + region.len()).align_up(0x8_0000_0000);
-                // SAFETY: The range is guaranteed not to access physical memory.
-                unsafe {
-                    IO_MEM_DISPATCHER
-                        .add_range(mmio_start_addr..(mmio_start_addr + HIGH_MMIO_SIZE));
-                }
-            }
-        } else if (region.base() + region.len()) < u32::MAX as usize {
-            // Find the TOLM (Top of Low Memory) and initialize Low MMIO region (TOLM ~ 0xffff_ffff).
-            if region.typ() == MemoryRegionType::Usable {
-                // Align start address to 0x1000_0000
-                let mmio_start_addr = (region.base() + region.len()).align_up(0x1000_0000);
-                // SAFETY: The range is guaranteed not to access physical memory.
-                unsafe {
-                    IO_MEM_DISPATCHER.add_range(mmio_start_addr..0x1_0000_0000);
-                }
-                break;
-            }
-        }
-    }
-    if !is_tohm_found {
-        let mmio_start_addr = 0x8_0000_0000;
-        // SAFETY: The range is guaranteed not to access physical memory.
-        unsafe {
-            IO_MEM_DISPATCHER.add_range(mmio_start_addr..(mmio_start_addr + HIGH_MMIO_SIZE));
-        }
-    }
+    init_io_mem_dispatcher(crate::boot::memory_regions().to_owned(), &IO_MEM_DISPATCHER);
 }
